Presenting... the marvelous geocoding proxy API!
This is the initial commit of the geocoding proxy API challenge as
required by PostMates. I will admit that as primitive as it is, it took me
most of the day to get this designed and built. It is fairly minimal, but
without someone on the phone answering questions, it wasn't clear that
embellishments were needed or even desirable.

Don't forget to fill in your API keys when playing with this.